fix(estimates): keep Excel parser out of Worker bundle
